working integration test
